Add documentation and add separate stdin_format test module
